requirements.txt
Django==4.2.7
redis==6.2.0
celery==5.5.3
pillow==11.3.0
reportlab==4.4.3
django-simple-history==3.10.1
python-decouple==3.8
psycopg2-binary==2.9.10
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0 